Initial flattened commit
This commit is based on the 4.4.0 release of an earlier private
repository containing a project with hardcoded sites and users.
It is the starting point of a new, orphaned branch from which
all previous sites, credentials and commits have been stripped
in order to ensure that future releases from this point on can
be made publicly available on GitHub.